Keep matcher work tests out of crosscheck generation
Refs #792
