+++ safere/src/test/java/org/safere/MatcherTest.java
@@ -14,12 +14,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
 
 import java.time.Duration;
-import java.util.function.Consumer;
-import java.util.function.IntConsumer;
 import java.util.regex.MatchResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -243,25 +240,6 @@ class MatcherTest {
     }
 
     @Test
-    @Tag("work-counter")
-    @DisabledForCrosscheck("java.util.regex backtracking makes this a SafeRE linear-time check")
-    @DisplayName("group access stays linear for ambiguous repeated captures")
-    void groupAccessWithAmbiguousRepeatedCapturesStaysLinear() {
-      Pattern p = Pattern.compile("((a|aa))*");
-
-      assertNoWorkCliff(
-          "matches()+group(1)",
-          "a".repeat(100),
-          "a".repeat(5_000),
-          input -> {
-            Matcher m = p.matcher(input);
-            assertThat(m.matches()).isTrue();
-            assertThat(m.group(1)).isEqualTo("a");
-            assertThat(m.group(2)).isEqualTo("a");
-          });
-    }
-
-    @Test
     @DisabledForCrosscheck("java.util.regex backtracking makes this a SafeRE linear-time check")
     @DisplayName("group access stays stack-safe for large repeated captures")
     void groupAccessWithLargeRepeatedCapturesStaysStackSafe() {
@@ -299,41 +277,6 @@ class MatcherTest {
       assertThat(m.group(0)).isEqualTo("123abc");
       assertThat(m.group(1)).isEqualTo("123");
       assertThat(m.group(2)).isEqualTo("abc");
-    }
-
-    @Test
-    @Tag("work-counter")
-    @DisabledForCrosscheck("java.util.regex backtracks on this SafeRE linear-time stress case")
-    @DisplayName("matches() stays linear for repeated dot-star with bounded captures")
-    void matchesWithRepeatedDotStarAndBoundedCaptures() {
-      Pattern p = repeatedDotStarSqlUnionPattern();
-
-      assertNoWorkCliff(
-          "matches()",
-          blocks -> {
-            String input = repeatedDotStarSqlUnionInput(blocks);
-            Matcher m = p.matcher(input);
-            assertThat(m.matches()).isTrue();
-            assertThat(m.group()).isEqualTo(input);
-            assertThat(m.start()).isEqualTo(0);
-            assertThat(m.end()).isEqualTo(input.length());
-          });
-    }
-
-    @Test
-    @Tag("work-counter")
-    @DisabledForCrosscheck("java.util.regex backtracks on this SafeRE linear-time stress case")
-    @DisplayName("lookingAt() stays linear for repeated dot-star with bounded captures")
-    void lookingAtWithRepeatedDotStarAndBoundedCaptures() {
-      Pattern p = repeatedDotStarSqlUnionPattern();
-
-      assertNoWorkCliff(
-          "lookingAt()",
-          blocks -> {
-            Matcher m = p.matcher(repeatedDotStarSqlUnionInput(blocks));
-            assertThat(m.lookingAt()).isTrue();
-            assertThat(m.group(1)).contains("INFORMATION_SCHEMA");
-          });
     }
 
     @ParameterizedTest
@@ -374,22 +317,6 @@ class MatcherTest {
   @Nested
   @DisplayName("find()")
   class FindTests {
-
-    @Test
-    @Tag("work-counter")
-    @DisabledForCrosscheck("java.util.regex backtracks on this SafeRE linear-time stress case")
-    @DisplayName("group access after find() stays linear for repeated dot-star captures")
-    void findGroupWithRepeatedDotStarAndBoundedCaptures() {
-      Pattern p = repeatedDotStarSqlUnionPattern();
-
-      assertNoWorkCliff(
-          "find()+group(1)",
-          blocks -> {
-            Matcher m = p.matcher(repeatedDotStarSqlUnionInput(blocks));
-            assertThat(m.find()).isTrue();
-            assertThat(m.group(1)).contains("INFORMATION_SCHEMA");
-          });
-    }
 
     @Test
     @DisplayName("find() locates a single match in the input")
@@ -2456,24 +2383,6 @@ class MatcherTest {
   class RegionTests {
 
     @Test
-    @Tag("work-counter")
-    @DisabledForCrosscheck("java.util.regex backtracks on this SafeRE linear-time stress case")
-    @DisplayName("region find stays linear for repeated dot-star captures")
-    void regionFindWithRepeatedDotStarAndBoundedCaptures() {
-      Pattern p = repeatedDotStarSqlUnionPattern();
-
-      assertNoWorkCliff(
-          "region().find()",
-          blocks -> {
-            String input = "prefix\n" + repeatedDotStarSqlUnionInput(blocks) + "suffix\n";
-            Matcher m = p.matcher(input);
-            m.region("prefix\n".length(), input.length() - "suffix\n".length());
-            assertThat(m.find()).isTrue();
-            assertThat(m.group(1)).contains("INFORMATION_SCHEMA");
-          });
-    }
-
-    @Test
     @DisplayName("find() respects region boundaries")
     void findRespectsRegion() {
       Pattern p = Pattern.compile("\\d+");
@@ -3270,59 +3179,6 @@ class MatcherTest {
     assertThat(m.find()).isTrue();
     assertThat(m.group()).isEmpty();
     assertThat(m.find()).isFalse();
-  }
-
-  private static String repeatedDotStarSqlUnionInput(int selectCount) {
-    StringBuilder input = new StringBuilder();
-    for (int i = 1; i <= selectCount; i++) {
-      input
-          .append("(SELECT *, PARSE_DATE('%Y-%m-%d', '2025-06-25') AS snapshot_date FROM ")
-          .append("`project-")
-          .append("%02d".formatted(i))
-          .append("`.`region2`.INFORMATION_SCHEMA.TABLE_OPTIONS)\n")
-          .append("UNION ALL\n");
-    }
-    return input.toString();
-  }
-
-  private static Pattern repeatedDotStarSqlUnionPattern() {
-    return Pattern.compile(
-        ".*SELECT.*FROM.*(.*INFORMATION_SCHEMA.*){5,}.*",
-        Pattern.CASE_INSENSITIVE | Pattern.DOTALL);
-  }
-
-  private static void assertNoWorkCliff(String api, IntConsumer scenario) {
-    long largerPositiveWork = countedWork(() -> scenario.accept(16));
-    long nearMinimumWork = countedWork(() -> scenario.accept(5));
-
-    assertThat(largerPositiveWork).as("%s should use matcher work", api).isPositive();
-    assertThat(nearMinimumWork)
-        .as(
-            "%s near-minimum input should not be dramatically slower than a larger "
-                + "positive input; nearWork=%d largerWork=%d",
-            api, nearMinimumWork, largerPositiveWork)
-        .isLessThan(largerPositiveWork * 50);
-  }
-
-  private static void assertNoWorkCliff(
-      String api, String nearMinimumInput, String largerPositiveInput, Consumer<String> scenario) {
-    scenario.accept(nearMinimumInput);
-    scenario.accept(largerPositiveInput);
-    long largerPositiveWork = countedWork(() -> scenario.accept(largerPositiveInput));
-    long nearMinimumWork = countedWork(() -> scenario.accept(nearMinimumInput));
-
-    assertThat(largerPositiveWork).as("%s should use matcher work", api).isPositive();
-    assertThat(nearMinimumWork)
-        .as(
-            "%s near-minimum input should not be dramatically slower than a larger "
-                + "positive input; nearWork=%d largerWork=%d",
-            api, nearMinimumWork, largerPositiveWork)
-        .isLessThan(largerPositiveWork * 50);
-  }
-
-  private static long countedWork(Runnable task) {
-    return assertTimeoutPreemptively(
-        PERFORMANCE_SCENARIO_TIMEOUT, () -> WorkCounter.countForTesting(task));
   }
 
   private static void assertZeroCountGroupBehavior(String regex, String input) {
